+++ src/Symfony/Component/Notifier/Bridge/AllMySms/Tests/AllMySmsTransportTest.php
@@ -13,9 +13,10 @@ namespace Symfony\Component\Notifier\Bridge\AllMySms\Tests;
 
 use Symfony\Component\Notifier\Bridge\AllMySms\AllMySmsTransport;
 use Symfony\Component\Notifier\Message\ChatMessage;
-use Symfony\Component\Notifier\Message\MessageInterface;
 use Symfony\Component\Notifier\Message\SmsMessage;
 use Symfony\Component\Notifier\Test\TransportTestCase;
+use Symfony\Component\Notifier\Tests\Fixtures\DummyHttpClient;
+use Symfony\Component\Notifier\Tests\Fixtures\DummyMessage;
 use Symfony\Component\Notifier\Transport\TransportInterface;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
@@ -24,25 +25,25 @@ final class AllMySmsTransportTest extends TransportTestCase
     /**
      * @return AllMySmsTransport
      */
-    public function createTransport(HttpClientInterface $client = null, string $from = null): TransportInterface
+    public static function createTransport(HttpClientInterface $client = null, string $from = null): TransportInterface
     {
-        return new AllMySmsTransport('login', 'apiKey', $from, $client ?? $this->createMock(HttpClientInterface::class));
+        return new AllMySmsTransport('login', 'apiKey', $from, $client ?? new DummyHttpClient());
     }
 
-    public function toStringProvider(): iterable
+    public static function toStringProvider(): iterable
     {
-        yield ['allmysms://api.allmysms.com', $this->createTransport()];
-        yield ['allmysms://api.allmysms.com?from=TEST', $this->createTransport(null, 'TEST')];
+        yield ['allmysms://api.allmysms.com', self::createTransport()];
+        yield ['allmysms://api.allmysms.com?from=TEST', self::createTransport(null, 'TEST')];
     }
 
-    public function supportedMessagesProvider(): iterable
+    public static function supportedMessagesProvider(): iterable
     {
         yield [new SmsMessage('0611223344', 'Hello!')];
     }
 
-    public function unsupportedMessagesProvider(): iterable
+    public static function unsupportedMessagesProvider(): iterable
     {
         yield [new ChatMessage('Hello!')];
-        yield [$this->createMock(MessageInterface::class)];
+        yield [new DummyMessage()];
     }
 }