+++ src/Symfony/Component/Notifier/Bridge/Clickatell/Tests/ClickatellTransportTest.php
@@ -19,6 +19,8 @@ use Symfony\Component\Notifier\Message\ChatMessage;
 use Symfony\Component\Notifier\Message\MessageInterface;
 use Symfony\Component\Notifier\Message\SmsMessage;
 use Symfony\Component\Notifier\Test\TransportTestCase;
+use Symfony\Component\Notifier\Tests\Fixtures\DummyHttpClient;
+use Symfony\Component\Notifier\Tests\Fixtures\DummyMessage;
 use Symfony\Component\Notifier\Transport\TransportInterface;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 use Symfony\Contracts\HttpClient\ResponseInterface;
@@ -28,31 +30,31 @@ final class ClickatellTransportTest extends TransportTestCase
     /**
      * @return ClickatellTransport
      */
-    public function createTransport(HttpClientInterface $client = null, string $from = null): TransportInterface
+    public static function createTransport(HttpClientInterface $client = null, string $from = null): TransportInterface
     {
-        return new ClickatellTransport('authToken', $from, $client ?? $this->createMock(HttpClientInterface::class));
+        return new ClickatellTransport('authToken', $from, $client ?? new DummyHttpClient());
     }
 
-    public function toStringProvider(): iterable
+    public static function toStringProvider(): iterable
     {
-        yield ['clickatell://api.clickatell.com', $this->createTransport()];
-        yield ['clickatell://api.clickatell.com?from=TEST', $this->createTransport(null, 'TEST')];
+        yield ['clickatell://api.clickatell.com', self::createTransport()];
+        yield ['clickatell://api.clickatell.com?from=TEST', self::createTransport(null, 'TEST')];
     }
 
-    public function supportedMessagesProvider(): iterable
+    public static function supportedMessagesProvider(): iterable
     {
         yield [new SmsMessage('+33612345678', 'Hello!')];
     }
 
-    public function unsupportedMessagesProvider(): iterable
+    public static function unsupportedMessagesProvider(): iterable
     {
         yield [new ChatMessage('Hello!')];
-        yield [$this->createMock(MessageInterface::class)];
+        yield [new DummyMessage()];
     }
 
     public function testExceptionIsThrownWhenNonMessageIsSend()
     {
-        $transport = $this->createTransport();
+        $transport = self::createTransport();
 
         $this->expectException(LogicException::class);
 
@@ -77,7 +79,7 @@ final class ClickatellTransportTest extends TransportTestCase
 
         $client = new MockHttpClient($response);
 
-        $transport = $this->createTransport($client);
+        $transport = self::createTransport($client);
 
         $this->expectException(TransportException::class);
         $this->expectExceptionMessage('Unable to send SMS with Clickatell: Error code 105 with message "Invalid Account Reference EX0000000" (https://documentation-page).');