+++ src/Symfony/Component/Notifier/Bridge/Expo/Tests/ExpoTransportTest.php
@@ -12,10 +12,11 @@
 namespace Symfony\Component\Notifier\Bridge\Expo\Tests;
 
 use Symfony\Component\Notifier\Bridge\Expo\ExpoTransport;
-use Symfony\Component\Notifier\Message\MessageInterface;
 use Symfony\Component\Notifier\Message\PushMessage;
 use Symfony\Component\Notifier\Message\SmsMessage;
 use Symfony\Component\Notifier\Test\TransportTestCase;
+use Symfony\Component\Notifier\Tests\Fixtures\DummyHttpClient;
+use Symfony\Component\Notifier\Tests\Fixtures\DummyMessage;
 use Symfony\Component\Notifier\Transport\TransportInterface;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
@@ -27,24 +28,24 @@ final class ExpoTransportTest extends TransportTestCase
     /**
      * @return ExpoTransport
      */
-    public function createTransport(HttpClientInterface $client = null): TransportInterface
+    public static function createTransport(HttpClientInterface $client = null): TransportInterface
     {
-        return new ExpoTransport('token', $client ?? $this->createMock(HttpClientInterface::class));
+        return new ExpoTransport('token', $client ?? new DummyHttpClient());
     }
 
-    public function toStringProvider(): iterable
+    public static function toStringProvider(): iterable
     {
-        yield ['expo://exp.host/--/api/v2/push/send', $this->createTransport()];
+        yield ['expo://exp.host/--/api/v2/push/send', self::createTransport()];
     }
 
-    public function supportedMessagesProvider(): iterable
+    public static function supportedMessagesProvider(): iterable
     {
         yield [new PushMessage('Hello!', 'Symfony Notifier')];
     }
 
-    public function unsupportedMessagesProvider(): iterable
+    public static function unsupportedMessagesProvider(): iterable
     {
         yield [new SmsMessage('0670802161', 'Hello!')];
-        yield [$this->createMock(MessageInterface::class)];
+        yield [new DummyMessage()];
     }
 }