+++ src/Symfony/Component/Notifier/Bridge/Sms77/Tests/Sms77TransportTest.php
@@ -16,6 +16,8 @@ use Symfony\Component\Notifier\Message\ChatMessage;
 use Symfony\Component\Notifier\Message\MessageInterface;
 use Symfony\Component\Notifier\Message\SmsMessage;
 use Symfony\Component\Notifier\Test\TransportTestCase;
+use Symfony\Component\Notifier\Tests\Fixtures\DummyHttpClient;
+use Symfony\Component\Notifier\Tests\Fixtures\DummyMessage;
 use Symfony\Component\Notifier\Transport\TransportInterface;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
@@ -24,25 +26,25 @@ final class Sms77TransportTest extends TransportTestCase
     /**
      * @return Sms77Transport
      */
-    public function createTransport(HttpClientInterface $client = null, string $from = null): TransportInterface
+    public static function createTransport(HttpClientInterface $client = null, string $from = null): TransportInterface
     {
-        return new Sms77Transport('apiKey', $from, $client ?? $this->createMock(HttpClientInterface::class));
+        return new Sms77Transport('apiKey', $from, $client ?? new DummyHttpClient());
     }
 
-    public function toStringProvider(): iterable
+    public static function toStringProvider(): iterable
     {
-        yield ['sms77://gateway.sms77.io', $this->createTransport()];
-        yield ['sms77://gateway.sms77.io?from=TEST', $this->createTransport(null, 'TEST')];
+        yield ['sms77://gateway.sms77.io', self::createTransport()];
+        yield ['sms77://gateway.sms77.io?from=TEST', self::createTransport(null, 'TEST')];
     }
 
-    public function supportedMessagesProvider(): iterable
+    public static function supportedMessagesProvider(): iterable
     {
         yield [new SmsMessage('0611223344', 'Hello!')];
     }
 
-    public function unsupportedMessagesProvider(): iterable
+    public static function unsupportedMessagesProvider(): iterable
     {
         yield [new ChatMessage('Hello!')];
-        yield [$this->createMock(MessageInterface::class)];
+        yield [new DummyMessage()];
     }
 }